move test case to another suite
